Rename _handler.py to _basehandler.py
