+++ telegram/ext/__init__.py
@@ -63,6 +63,7 @@ from . import filters
 from ._aioratelimiter import AIORateLimiter
 from ._application import Application, ApplicationHandlerStop
 from ._applicationbuilder import ApplicationBuilder
+from ._basehandler import BaseHandler
 from ._basepersistence import BasePersistence, PersistenceInput
 from ._baseratelimiter import BaseRateLimiter
 from ._baseupdateprocessor import BaseUpdateProcessor, SimpleUpdateProcessor
@@ -78,7 +79,6 @@ from ._conversationhandler import ConversationHandler
 from ._defaults import Defaults
 from ._dictpersistence import DictPersistence
 from ._extbot import ExtBot
-from ._handler import BaseHandler
 from ._inlinequeryhandler import InlineQueryHandler
 from ._jobqueue import Job, JobQueue
 from ._messagehandler import MessageHandler
