+++ telegram/ext/_commandhandler.py
@@ -24,7 +24,7 @@ from telegram import MessageEntity, Update
 from telegram._utils.defaultvalue import DEFAULT_TRUE
 from telegram._utils.types import SCT, DVType
 from telegram.ext import filters as filters_module
-from telegram.ext._handler import BaseHandler
+from telegram.ext._basehandler import BaseHandler
 from telegram.ext._utils.types import CCT, FilterDataDict, HandlerCallback
 
 if TYPE_CHECKING:
