+++ telegram/ext/_conversationhandler.py
@@ -41,10 +41,10 @@ from telegram._utils.logging import get_logger
 from telegram._utils.types import DVType
 from telegram._utils.warnings import warn
 from telegram.ext._application import ApplicationHandlerStop
+from telegram.ext._basehandler import BaseHandler
 from telegram.ext._callbackqueryhandler import CallbackQueryHandler
 from telegram.ext._choseninlineresulthandler import ChosenInlineResultHandler
 from telegram.ext._extbot import ExtBot
-from telegram.ext._handler import BaseHandler
 from telegram.ext._inlinequeryhandler import InlineQueryHandler
 from telegram.ext._stringcommandhandler import StringCommandHandler
 from telegram.ext._stringregexhandler import StringRegexHandler
