+++ telegram/ext/_stringcommandhandler.py
@@ -22,7 +22,7 @@ from typing import TYPE_CHECKING, Any, List, Optional
 
 from telegram._utils.defaultvalue import DEFAULT_TRUE
 from telegram._utils.types import DVType
-from telegram.ext._handler import BaseHandler
+from telegram.ext._basehandler import BaseHandler
 from telegram.ext._utils.types import CCT, RT, HandlerCallback
 
 if TYPE_CHECKING:
